MIPS64: Fix `[callbacks] Properly support RCS for ApiAccessors and Callbacks.`
Fix 0027c83440059e6717032ff2cd1ba677eb9890e7

FLAG_runtime_stats was declared as Int but it was accessed as
pointer. On 64-bit architectures pointer access needs to be
eight byte aligned. This created unaligned memory access
on 64-bit architectures for those architectures that do not
support unaligned memory access.

TEST=mjsunit/wasm/table,cctest/test-accessors/AccessorIC

+++ src/code-stub-assembler.cc
@@ -12086,7 +12086,7 @@ Node* CodeStubAssembler::IsDebugActive() {
 
 TNode<BoolT> CodeStubAssembler::IsRuntimeCallStatsEnabled() {
   TNode<UintPtrT> flag_value = UncheckedCast<UintPtrT>(Load(
-      MachineType::IntPtr(),
+      MachineType::Int32(),
       ExternalConstant(ExternalReference::address_of_runtime_stats_flag())));
   return WordNotEqual(flag_value, IntPtrConstant(0));
 }
